Added check if the coverage file exists.

+++ coverage/run_coverage.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-rm .coverage
+if [ -e ".coverage" ]; then
+    rm .coverage
+fi
+
 python-coverage run ../test/run_tests.py
 python-coverage html
 firefox htmlcov/index.html
